SCP-2709: Refactoring code structure
* Renaming Definitions to Domain
