+++ marlowe-actus/src/Language/Marlowe/ACTUS/Model/STF/StateTransitionModel.hs
@@ -3,15 +3,15 @@
 
 module Language.Marlowe.ACTUS.Model.STF.StateTransitionModel where
 
-import           Data.Maybe                                        (fromMaybe)
-import           Language.Marlowe.ACTUS.Definitions.BusinessEvents (RiskFactorsPoly (..))
-import           Language.Marlowe.ACTUS.Definitions.ContractState  (ContractStatePoly (..))
-import           Language.Marlowe.ACTUS.Definitions.ContractTerms  (ContractTermsPoly (..), FEB (..), IPCB (..),
-                                                                    OPTP (..), SCEF (..))
-import           Language.Marlowe.ACTUS.Definitions.Ops            (ActusNum (..), ActusOps (..), DateOps (_lt),
-                                                                    RoleSignOps (_r))
-import           Language.Marlowe.ACTUS.Utility.ANN.Annuity        (annuity)
-import           Prelude                                           hiding (Fractional, Num, (*), (+), (-), (/))
+import           Data.Maybe                                   (fromMaybe)
+import           Language.Marlowe.ACTUS.Domain.BusinessEvents (RiskFactorsPoly (..))
+import           Language.Marlowe.ACTUS.Domain.ContractState  (ContractStatePoly (..))
+import           Language.Marlowe.ACTUS.Domain.ContractTerms  (ContractTermsPoly (..), FEB (..), IPCB (..), OPTP (..),
+                                                               SCEF (..))
+import           Language.Marlowe.ACTUS.Domain.Ops            (ActusNum (..), ActusOps (..), DateOps (_lt),
+                                                               RoleSignOps (_r))
+import           Language.Marlowe.ACTUS.Utility.ANN.Annuity   (annuity)
+import           Prelude                                      hiding (Fractional, Num, (*), (+), (-), (/))
 
 -- Principal at Maturity (PAM)
