+++ marlowe-actus/src/Language/Marlowe/ACTUS/Utility/YearFraction.hs
@@ -2,10 +2,10 @@
 
 module Language.Marlowe.ACTUS.Utility.YearFraction where
 
-import           Data.Time                                        (Day, LocalTime (..), TimeOfDay (..), addLocalTime,
-                                                                   diffDays, fromGregorian, gregorianMonthLength,
-                                                                   isLeapYear, toGregorian)
-import           Language.Marlowe.ACTUS.Definitions.ContractTerms (DCC (DCC_A_360, DCC_A_365, DCC_A_AISDA, DCC_E30_360, DCC_E30_360ISDA))
+import           Data.Time                                   (Day, LocalTime (..), TimeOfDay (..), addLocalTime,
+                                                              diffDays, fromGregorian, gregorianMonthLength, isLeapYear,
+                                                              toGregorian)
+import           Language.Marlowe.ACTUS.Domain.ContractTerms (DCC (DCC_A_360, DCC_A_365, DCC_A_AISDA, DCC_E30_360, DCC_E30_360ISDA))
 
 yearFraction :: DCC -> LocalTime -> LocalTime -> Maybe LocalTime -> Double
 yearFraction dcc x y o = yearFraction' dcc (localDay x) (localDay $ clipToMidnight y) (localDay <$> o)
