+++ marlowe-actus/test/Spec/Marlowe/ACTUS/QCGenerator.hs
@@ -8,15 +8,15 @@
 
 module Spec.Marlowe.ACTUS.QCGenerator where
 
-import qualified Data.List                                         as L
-import qualified Data.Map                                          as M
-import           Data.Maybe                                        (catMaybes)
+import qualified Data.List                                    as L
+import qualified Data.Map                                     as M
+import           Data.Maybe                                   (catMaybes)
 import           Data.Time
-import           Data.Time.Clock.POSIX                             (posixSecondsToUTCTime)
-import           Data.Time.Clock.System                            (SystemTime (MkSystemTime), utcToSystemTime)
-import           Language.Marlowe.ACTUS.Definitions.BusinessEvents
-import           Language.Marlowe.ACTUS.Definitions.ContractTerms
-import           Language.Marlowe.ACTUS.Definitions.Schedule
+import           Data.Time.Clock.POSIX                        (posixSecondsToUTCTime)
+import           Data.Time.Clock.System                       (SystemTime (MkSystemTime), utcToSystemTime)
+import           Language.Marlowe.ACTUS.Domain.BusinessEvents
+import           Language.Marlowe.ACTUS.Domain.ContractTerms
+import           Language.Marlowe.ACTUS.Domain.Schedule
 import           Language.Marlowe.ACTUS.Generator.Analysis
 import           Test.QuickCheck
